pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "airo-ipc"
version = "0.1.2"
authors = [
    { name = "Matthijs Biondina", email = "matthijs.biondina@ugent.be" },
    { name = "Mathieu De Coster", email = "mathieu.decoster@ugent.be" },
]
description = "IPC framework for AIRO."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "cyclonedds-nightly==2025.6.6",
    "loguru>=0.7.2",
    "numpy",
]

[project.urls]
Homepage = "https://github.com/airo-ugent/airo-ipc"
Issues = "https://github.com/airo-ugent/airo-ipc/issues"

[tools.setuptools]
packages = ["airo_ipc"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
